Implemented collision event collector
Implemented a event collector that only processes collision events until
it has been discussed on the next meeting whether or not the solution is
proper.

Major changes:
-Added tickCount as a long to GameData.
-Added incrementer and getter methods for tickCount.
-Added expirationTick to CollisionEvent.
-Edited all methods affected by the new parameter on CollisionEvent.

How:
-When the CollisionData is queried with the 'getEvents' method a
parameter is passed along with the call, this is the currentTick from
GameData. If the currentTick is the same as the last tick that was
queried for then it just returns the List of events as before. If the
currentTick is not the same as the last tick however, the list is
filtered, saved and then returned.