Implemented score functionality
The score is currently just a kill count stored in the Player entity
defined by PlayerCommon. For this to work I also added mob death event
to the removal of the mob itself in the MobSystem and changed the
MobData to use the new Event 'system' which will be refactored soon for
easier use and maintainability.