Re-re-implemented fallback texture
Re-re-implemented the fallback texture because the most recent commit
got rid of it while also making the entire thing crash for me at least
and going by what the fallback texture was applied to it has something
to do with the background? Also, please stop removing the fallback
texture functionality.