Issue resolving and refactoring
The last commit caused a new issue as described in #3 on the issues
page. This has now been resolved by piggybacking off of the
stackoverflow link also provided in #3. Specifically the dispose event
has been taken into use to solve the issue, in the future we should
avoid removing entities in the midst of the update cycle. (The dispose
event is posted by the game class at the end of the update cycle).

Aside from that the world class has been modified.
-getEntitiesByClass now handles regular events aswell as anonymous
events.
-getEntitiesByClass is now capable of returning a list of any class that
extends entity, this means much less casting is needed in the handlers
and that instanceof checks can be avoided.
-removeEntities has been added, it takes a collection of any class that
extends entity and removes every entity in the list from the world.

Some of the code has already been refactored so that the above mentioned
methods are utilized but the source is extensive so it is nowhere near
complete.