New subscribe / unsubscribe methods plus refactoring
Added new subscribe method which discovers all eventhandlers declared in
a class reflectively, this makes the use of the dispatcher way easier
and way more convenient. A similar unsubscribe method has also been
added.

Added constructors to the pair structure, one to initialize the pair
with argument objects and one which initializes both fields to null so
that 'legacy' code is still functional.

Had problem with powerup not being removed reliably, resolved by
restricting the clearing of bad entity lists (lists of those entities
that are about to be removed) to the disposal handlers.